refactor(cart): implement guest user handling in add to cart functionality
