+++ lib/core/utils/shared_features/add_to_cart/data/data_source/contract/add_to_cart_local_data_source.dart
@@ -1,1 +1,3 @@
-abstract interface class AddToCartLocalDataSource {}
+abstract interface class AddToCartLocalDataSource {
+  Future<String?> isGuestUser();
+}
