+++ lib/core/utils/shared_features/add_to_cart/data/repo_impl/add_to_cart_repo_impl.dart
@@ -3,6 +3,7 @@ import 'package:ecommerce_flower_app/core/utils/datasource_excution/api_result.d
 import 'package:injectable/injectable.dart';
 
 import '../../domain/repo/add_to_cart_repo.dart';
+import '../data_source/contract/add_to_cart_local_data_source.dart';
 import '../data_source/contract/add_to_cart_remote_data_source.dart';
 import '../model/request/add_to_cart_request_dto.dart';
 import '../model/response/add_to_cart_response_dto.dart';
@@ -11,7 +12,12 @@ import '../model/response/add_to_cart_response_dto.dart';
 class AddToCartRepoImpl implements AddToCartRepo {
   final ApiManager _apiManager;
   final AddToCartRemoteDataSource _addToCartRemoteDataSource;
-  AddToCartRepoImpl(this._apiManager, this._addToCartRemoteDataSource);
+  final AddToCartLocalDataSource _addToCartLocalDataSource;
+  AddToCartRepoImpl(
+    this._apiManager,
+    this._addToCartRemoteDataSource,
+    this._addToCartLocalDataSource,
+  );
   @override
   Future<Result<AddToCartResponseDto>> addToCart(
     AddToCartRequestDto request,
@@ -22,5 +28,10 @@ class AddToCartRepoImpl implements AddToCartRepo {
       );
       return response;
     });
+  }
+
+  @override
+  Future<String?> isGuestUser() async {
+    return await _addToCartLocalDataSource.isGuestUser();
   }
 }
