+++ lib/core/utils/shared_features/add_to_cart/domain/usecase/is_guest_use_case.dart
@@ -1,0 +1,13 @@
+import 'package:injectable/injectable.dart';
+
+import '../repo/add_to_cart_repo.dart';
+
+@injectable
+class IsGuestUseCase {
+  final AddToCartRepo _addToCartRepo;
+  IsGuestUseCase(this._addToCartRepo);
+
+  Future<String?> call() async {
+    return await _addToCartRepo.isGuestUser();
+  }
+}
