--- lib/core/utils/shared_features/is_guest/data/data_source/contract/is_guest_local_data_source.dart
@@ -1,3 +1,0 @@
-abstract interface class IsGuestLocalDataSource {
-  Future<String?> isGuestUser();
-}
