--- lib/core/utils/shared_features/is_guest/data/data_source/contract/is_guest_remote_data_source.dart
@@ -1,1 +1,0 @@
-abstract interface class IsGuestRemoteDataSource {}
