--- lib/core/utils/shared_features/is_guest/data/data_source/remote/is_guest_remote_data_source_impl.dart
@@ -1,6 +1,0 @@
-import 'package:injectable/injectable.dart';
-
-import '../contract/is_guest_remote_data_source.dart';
-
-@Injectable(as: IsGuestRemoteDataSource)
-class IsGuestRemoteDataSourceImpl implements IsGuestRemoteDataSource {}
