--- lib/core/utils/shared_features/is_guest/domain/repo/is_guest_repo.dart
@@ -1,3 +1,0 @@
-abstract interface class IsGuestRepo {
-  Future<String?> isGuestUser();
-}
