--- lib/core/utils/shared_features/is_guest/domain/usecase/is_guest_use_case.dart
@@ -1,13 +1,0 @@
-import 'package:injectable/injectable.dart';
-
-import '../repo/is_guest_repo.dart';
-
-@injectable
-class IsGuestUseCase {
-  final IsGuestRepo _isGuestRepo;
-  IsGuestUseCase(this._isGuestRepo);
-
-  Future<String?> call() async {
-    return await _isGuestRepo.isGuestUser();
-  }
-}
