--- lib/core/utils/shared_features/is_guest/presentation/view/guest_mode_section.dart
@@ -1,40 +1,0 @@
-import 'package:easy_localization/easy_localization.dart';
-import 'package:flutter/material.dart';
-
-import '../../../../l10n/locale_keys.g.dart';
-import '../../../../routes/routes.dart';
-
-class GuestModeSection extends StatelessWidget {
-  const GuestModeSection({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Text(
-            LocaleKeys.YouAreInGuestMode.tr(),
-            style: const TextStyle(fontSize: 20),
-            textAlign: TextAlign.center,
-          ),
-          const SizedBox(height: 16),
-          Text(
-            LocaleKeys.PleaseLoginToAccessAllFeatures.tr(),
-            style: const TextStyle(fontSize: 20),
-            textAlign: TextAlign.center,
-          ),
-          const SizedBox(height: 32),
-          ElevatedButton(
-            onPressed: () {
-              Navigator.pushReplacementNamed(context, AppRoutes.loginRoute);
-            },
-            child: Text(LocaleKeys.Login.tr()),
-          ),
-        ],
-      ),
-    );
-  }
-}
